+++ setup.py
@@ -24,7 +24,7 @@ XBLOCKS = [
     "library_content = xmodule.library_content_block:LibraryContentBlock",
     "library_sourced = xmodule.library_sourced_block:LibrarySourcedBlock",
     "lti = xmodule.lti_block:LTIBlock",
-    "poll_question = xmodule.poll_module:PollBlock",
+    "poll_question = xmodule.poll_block:PollBlock",
     "problem = xmodule.capa_block:ProblemBlock",
     "randomize = xmodule.randomize_module:RandomizeBlock",
     "sequential = xmodule.seq_module:SequenceBlock",
